uhd: allow for 16 bit i2c and eeprom addrs
Conflicts:
	host/lib/usrp/b250/b250_dboard_iface.cpp
	host/lib/usrp/cores/i2c_core_100_wb32.cpp

+++ host/include/uhd/types/serial.hpp
@@ -57,7 +57,7 @@ namespace uhd{
          * \param buf the vector of bytes
          */
         virtual void write_i2c(
-            boost::uint8_t addr,
+            boost::uint16_t addr,
             const byte_vector_t &buf
         ) = 0;
 
@@ -68,7 +68,7 @@ namespace uhd{
          * \return a vector of bytes
          */
         virtual byte_vector_t read_i2c(
-            boost::uint8_t addr,
+            boost::uint16_t addr,
             size_t num_bytes
         ) = 0;
 
@@ -79,8 +79,8 @@ namespace uhd{
          * \param buf the vector of bytes
          */
         virtual void write_eeprom(
-            boost::uint8_t addr,
-            boost::uint8_t offset,
+            boost::uint16_t addr,
+            boost::uint16_t offset,
             const byte_vector_t &buf
         );
 
@@ -92,8 +92,8 @@ namespace uhd{
          * \return a vector of bytes
          */
         virtual byte_vector_t read_eeprom(
-            boost::uint8_t addr,
-            boost::uint8_t offset,
+            boost::uint16_t addr,
+            boost::uint16_t offset,
             size_t num_bytes
         );
     };
